# openapi7.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Hamdaz ERP",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/v1/users/{ref}": {
   "get": {
    "tags": [
     "users"
    ],
    "summary": "Everything about one user",
    "operationId": "get_profile_api_v1_users__ref__get",
    "parameters": [
     {
      "name": "ref",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Ref"
      }
     },
     {
      "name": "include",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "Comma-separated section keys. Omit for all of them.",
       "title": "Include"
      },
      "description": "Comma-separated section keys. Omit for all of them."
     },
     {
      "name": "local_only",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "description": "Skip sections that call out to Entra",
       "default": false,
       "title": "Local Only"
      },
      "description": "Skip sections that call out to Entra"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UserProfileOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "users"
    ],
    "summary": "Remove the user from the ERP entirely",
    "operationId": "purge_user_api_v1_users__ref__delete",
    "parameters": [
     {
      "name": "ref",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Ref"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ResetResult"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/users/{ref}/reset": {
   "post": {
    "tags": [
     "users"
    ],
    "summary": "Strip every module's data, keeping the account",
    "operationId": "reset_user_api_v1_users__ref__reset_post",
    "parameters": [
     {
      "name": "ref",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Ref"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ResetResult"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/proposals/my-tasks": {
   "get": {
    "tags": [
     "proposals"
    ],
    "summary": "The proposal tasks assigned to the caller",
    "operationId": "my_tasks_api_v1_proposals_my_tasks_get",
    "parameters": [
     {
      "name": "open_only",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "description": "Hide tasks whose status is Completed",
       "default": true,
       "title": "Open Only"
      },
      "description": "Hide tasks whose status is Completed"
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 500,
       "minimum": 1,
       "default": 200,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/MyTasksOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/proposals/workload": {
   "get": {
    "tags": [
     "proposals"
    ],
    "summary": "Per-person proposal counts, for admins",
    "description": "Every person's totals in one call.\n\nAdmin-only, and deliberately not gated on the proposals *module*: this is an\norganisation-wide management view, not the team-scoped worker view that\n``/my-tasks`` serves.",
    "operationId": "workload_api_v1_proposals_workload_get",
    "parameters": [
     {
      "name": "team",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "Restrict to one team's members. Omit for everyone.",
       "title": "Team"
      },
      "description": "Restrict to one team's members. Omit for everyone."
     },
     {
      "name": "refresh",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "description": "Re-sweep SharePoint instead of using the cache",
       "default": false,
       "title": "Refresh"
      },
      "description": "Re-sweep SharePoint instead of using the cache"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/WorkloadOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "ExcludedOut": {
    "properties": {
     "rows": {
      "type": "integer",
      "title": "Rows"
     },
     "people": {
      "type": "integer",
      "title": "People"
     },
     "names": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Names"
     }
    },
    "type": "object",
    "required": [
     "rows",
     "people",
     "names"
    ],
    "title": "ExcludedOut",
    "description": "What a team scope left out. Never silently dropped."
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "MyTasksOut": {
    "properties": {
     "email": {
      "type": "string",
      "title": "Email"
     },
     "sharepoint_user_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Sharepoint User Id"
     },
     "in_sharepoint": {
      "type": "boolean",
      "title": "In Sharepoint"
     },
     "total": {
      "type": "integer",
      "title": "Total"
     },
     "open_count": {
      "type": "integer",
      "title": "Open Count"
     },
     "tasks": {
      "items": {
       "$ref": "#/components/schemas/TaskOut"
      },
      "type": "array",
      "title": "Tasks"
     }
    },
    "type": "object",
    "required": [
     "email",
     "sharepoint_user_id",
     "in_sharepoint",
     "total",
     "open_count",
     "tasks"
    ],
    "title": "MyTasksOut"
   },
...
   "PersonWorkloadOut": {
    "properties": {
     "lookup_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Lookup Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "email": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Email"
     },
     "total": {
      "type": "integer",
      "title": "Total"
     },
     "completed": {
      "type": "integer",
      "title": "Completed"
     },
     "open": {
      "type": "integer",
      "title": "Open"
     },
     "overdue": {
      "type": "integer",
      "title": "Overdue"
     },
     "due_soon": {
      "type": "integer",
      "title": "Due Soon"
     },
     "later": {
      "type": "integer",
      "title": "Later"
     },
     "no_deadline": {
      "type": "integer",
      "title": "No Deadline"
     },
     "no_status": {
      "type": "integer",
      "title": "No Status"
     },
     "by_status": {
      "additionalProperties": {
       "type": "integer"
      },
      "type": "object",
      "title": "By Status"
     },
     "next_deadline": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Next Deadline"
     }
    },
    "type": "object",
    "required": [
     "lookup_id",
     "name",
     "email",
     "total",
     "completed",
     "open",
     "overdue",
     "due_soon",
     "later",
     "no_deadline",
     "no_status",
     "by_status",
     "next_deadline"
    ],
    "title": "PersonWorkloadOut"
   },
...
   "ProfileMeta": {
    "properties": {
     "requested": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Requested"
     },
     "elapsed_ms": {
      "type": "integer",
      "title": "Elapsed Ms"
     },
     "section_ms": {
      "additionalProperties": {
       "type": "integer"
      },
      "type": "object",
      "title": "Section Ms"
     }
    },
    "type": "object",
    "required": [
     "requested",
     "elapsed_ms",
     "section_ms"
    ],
    "title": "ProfileMeta"
   },
...
   "ResetResult": {
    "properties": {
     "user_id": {
      "type": "string",
      "title": "User Id"
     },
     "email": {
      "type": "string",
      "title": "Email"
     },
     "display_name": {
      "type": "string",
      "title": "Display Name"
     },
     "account_deleted": {
      "type": "boolean",
      "title": "Account Deleted"
     },
     "removed": {
      "additionalProperties": {
       "type": "integer"
      },
      "type": "object",
      "title": "Removed"
     },
     "previous": {
      "additionalProperties": true,
      "type": "object",
      "title": "Previous"
     }
    },
    "type": "object",
    "required": [
     "user_id",
     "email",
     "display_name",
     "account_deleted",
     "removed",
     "previous"
    ],
    "title": "ResetResult"
   },
...
   "ScopeOut": {
    "properties": {
     "team_slug": {
      "type": "string",
      "title": "Team Slug"
     },
     "team_name": {
      "type": "string",
      "title": "Team Name"
     },
     "member_count": {
      "type": "integer",
      "title": "Member Count"
     },
     "matched_in_sharepoint": {
      "type": "integer",
      "title": "Matched In Sharepoint"
     },
     "members_without_sharepoint": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Members Without Sharepoint"
     }
    },
    "type": "object",
    "required": [
     "team_slug",
     "team_name",
     "member_count",
     "matched_in_sharepoint",
     "members_without_sharepoint"
    ],
    "title": "ScopeOut"
   },
...
   "TaskOut": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "title": {
      "type": "string",
      "title": "Title"
     },
     "status": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Status"
     },
     "priority": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Priority"
     },
     "assigned_to_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Assigned To Name"
     },
     "start_date": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Start Date"
     },
     "due_date": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Due Date"
     },
     "bid_closing_date": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Bid Closing Date"
     },
     "end_user": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "End User"
     },
     "submission_status": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Submission Status"
     },
     "current_type": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Current Type"
     },
     "order_status": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Order Status"
     },
     "negotiation": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Negotiation"
     },
     "quote_no": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Quote No"
     },
     "remarks": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Remarks"
     },
     "working_notes": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Working Notes"
     },
     "created_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Created At"
     },
     "modified_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Modified At"
     },
     "web_url": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Web Url"
     },
     "is_open": {
      "type": "boolean",
      "title": "Is Open"
     },
     "deadline": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Deadline"
     }
    },
    "type": "object",
    "required": [
     "id",
     "title",
     "status",
     "priority",
     "assigned_to_name",
     "start_date",
     "due_date",
     "bid_closing_date",
     "end_user",
     "submission_status",
     "current_type",
     "order_status",
     "negotiation",
     "quote_no",
     "remarks",
     "working_notes",
     "created_at",
     "modified_at",
     "web_url",
     "is_open",
     "deadline"
    ],
    "title": "TaskOut"
   },
...
   "UserProfileOut": {
    "properties": {
     "user_id": {
      "type": "string",
      "title": "User Id"
     },
     "email": {
      "type": "string",
      "title": "Email"
     },
     "display_name": {
      "type": "string",
      "title": "Display Name"
     },
     "sections": {
      "additionalProperties": true,
      "type": "object",
      "title": "Sections"
     },
     "errors": {
      "additionalProperties": {
       "type": "string"
      },
      "type": "object",
      "title": "Errors"
     },
     "meta": {
      "$ref": "#/components/schemas/ProfileMeta"
     }
    },
    "type": "object",
    "required": [
     "user_id",
     "email",
     "display_name",
     "sections",
     "errors",
     "meta"
    ],
    "title": "UserProfileOut"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
   "WorkloadOut": {
    "properties": {
     "organisation": {
      "$ref": "#/components/schemas/PersonWorkloadOut"
     },
     "people": {
      "items": {
       "$ref": "#/components/schemas/PersonWorkloadOut"
      },
      "type": "array",
      "title": "People"
     },
     "person_count": {
      "type": "integer",
      "title": "Person Count"
     },
     "soon_days": {
      "type": "integer",
      "title": "Soon Days"
     },
     "generated_at": {
      "type": "string",
      "title": "Generated At"
     },
     "row_count": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Row Count"
     },
     "scope": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/ScopeOut"
       },
       {
        "type": "null"
       }
      ]
     },
     "excluded": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/ExcludedOut"
       },
       {
        "type": "null"
       }
      ]
     },
     "cached": {
      "type": "boolean",
      "title": "Cached",
      "default": false
     },
     "age_seconds": {
      "type": "integer",
      "title": "Age Seconds",
      "default": 0
     },
     "elapsed_ms": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Elapsed Ms"
     },
     "fetch_ms": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Fetch Ms"
     }
    },
    "type": "object",
    "required": [
     "organisation",
     "people",
     "person_count",
     "soon_days",
     "generated_at"
    ],
    "title": "WorkloadOut"
   },
...
  }
 }
}